requirements.txt
absl-py==1.4.0
aiofiles==0.8.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asgiref==3.6.0
av==10.0.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachetools==5.3.0
certifi
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
fastapi==0.78.0
filelock==3.12.0
Flask==2.3.1
fonttools==4.39.3
fvcore==0.1.5.post20221221
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h11==0.14.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
loguru==0.6.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==1.0.0
ninja==1.11.1
numpy==1.22.4
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.5.64
packaging==23.1
pathspec==0.11.1
Pillow==9.0.1
platformdirs==3.5.0
portalocker==2.7.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydantic==1.9.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.19.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
tomli==2.0.1
torch==1.13.0
torchgeometry==0.1.2
torchvision==0.14.0
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.17.6
Werkzeug==2.3.2
yacs==0.1.8
zipp==3.15.0
